# openapi/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "OwlMail HTTP API",
    "version": "0.6.0",
    "summary": "Versioned OwlMail mailbox, relay, settings, and operational API",
    "description": "The contract for OwlMail's native `/api/v1` interface. Unversioned MailDev-style compatibility routes are intentionally excluded because their protocol behavior is not an exact MailDev contract. Configure `-base-pathname` (or `OWLMAIL_BASE_PATHNAME`) to place the server and this contract below a reverse-proxy prefix. The served documents rewrite the relative server URL to include that prefix. HTTP Basic authentication is optional at deployment time; when enabled it protects every operation except liveness and readiness. The unversioned `/healthz` and `/readyz` probes are also public but remain outside this versioned contract."
  },
...
  "servers": [
    {
      "url": "/api/v1",
      "description": "Root deployment; the served contract includes the configured base pathname"
    }
  ],
  "security": [
    {
      "basicAuth": []
    },
    {}
  ],
...
  "paths": {
...
    "/emails/{id}": {
      "get": {
        "tags": [
          "Emails"
        ],
        "summary": "Get one full email",
        "description": "Fetching details does not mark the email as read.",
        "operationId": "getEmail",
        "parameters": [
          {
            "$ref": "#/components/parameters/EmailID"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Email"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        }
      },
      "delete": {
        "tags": [
          "Emails"
        ],
        "summary": "Delete one email",
        "operationId": "deleteEmail",
        "parameters": [
          {
            "$ref": "#/components/parameters/EmailID"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIResponse"
                },
                "example": {
                  "code": "EMAIL_DELETED",
                  "message": "Email deleted"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        }
      }
    },
...
    "/emails/{id}/actions/relay/preflight": {
      "get": {
        "tags": [
          "Relay"
        ],
        "summary": "Preview effective relay recipients",
        "description": "Returns the current rule-filtered SMTP envelope recipients. Pass this exact list as confirmedRecipients when submitting the relay so configuration changes cannot expand delivery beyond the operator's confirmation.",
        "operationId": "relayEmailPreflight",
        "parameters": [
          {
            "$ref": "#/components/parameters/EmailID"
          }
        ],
        "responses": {
          "200": {
            "description": "Effective recipients for operator confirmation.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        }
      }
    },
...
    "/relay-jobs/{jobID}": {
      "get": {
        "tags": [
          "Relay"
        ],
        "summary": "Get asynchronous relay status",
        "description": "Returns a queryable relay job. With a persistent mail directory configured, queued jobs are durable and recovered after restart with at-least-once delivery semantics; the default process-specific temporary directory provides only same-process status. `attempts` is always present and `nextAttemptAt` appears while a retry is scheduled. Completed jobs expire after 24 hours, but the 1000-job capacity may evict them earlier after a one-minute post-completion protection window. Active jobs are never evicted.",
        "operationId": "getRelayJob",
        "parameters": [
          {
            "$ref": "#/components/parameters/RelayJobID"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIResponse"
                },
                "example": {
                  "code": "RELAY_STATUS",
                  "message": "Relay job status",
                  "data": {
                    "id": "0123456789abcdef0123456789abcdef",
                    "emailId": "aB3dEfGh",
                    "status": "queued",
                    "attempts": 1,
                    "createdAt": "2026-09-02T12:00:00Z",
                    "updatedAt": "2026-09-02T12:00:01Z",
                    "nextAttemptAt": "2026-09-02T12:00:02Z"
                  }
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "basicAuth": {
        "type": "http",
        "scheme": "basic",
        "description": "Deployment-conditional HTTP Basic Auth. When server credentials are configured, callers must supply matching credentials; rejected requests receive a plain-text 401 with `WWW-Authenticate: Basic realm=\"OwlMail\"`."
      }
    },
    "parameters": {
      "EmailID": {
        "name": "id",
        "in": "path",
        "required": true,
        "description": "Eight-character OwlMail ID or UUID, depending on server configuration.",
        "schema": {
          "type": "string",
          "minLength": 1
        },
        "example": "aB3dEfGh"
      },
...
      "RelayJobID": {
        "name": "jobID",
        "in": "path",
        "required": true,
        "description": "Opaque relay job identifier retained across process restarts when a persistent mail directory is configured.",
        "schema": {
          "type": "string",
          "minLength": 1
        },
        "example": "0123456789abcdef0123456789abcdef"
      },
...
    },
    "schemas": {
      "Address": {
        "type": "object",
        "description": "Address fields retain the capitalization produced by Go's mail.Address JSON encoding.",
        "required": [
          "Name",
          "Address"
        ],
        "properties": {
          "Name": {
            "type": "string",
            "examples": [
              "Sender"
            ]
          },
          "Address": {
            "type": "string",
            "examples": [
              "sender@example.com"
            ]
          }
        },
        "additionalProperties": false
      },
      "Attachment": {
        "type": "object",
        "required": [
          "contentType",
          "fileName",
          "generatedFileName",
          "contentId",
          "size"
        ],
        "properties": {
          "contentType": {
            "type": "string",
            "examples": [
              "application/pdf"
            ]
          },
          "fileName": {
            "type": "string",
            "examples": [
              "invoice.pdf"
            ]
          },
          "generatedFileName": {
            "type": "string",
            "examples": [
              "invoice.pdf"
            ]
          },
          "contentId": {
            "type": "string",
            "examples": [
              "logo@example.com"
            ]
          },
          "size": {
            "type": "integer",
            "format": "int64",
            "minimum": 0,
            "examples": [
              1280
            ]
          }
        },
        "additionalProperties": false
      },
...
      "Email": {
        "type": "object",
        "required": [
          "id",
          "time",
          "read",
          "subject",
          "from",
          "to",
          "cc",
          "bcc",
          "calculatedBcc",
          "text",
          "html",
          "attachments",
          "envelope",
          "source",
          "size",
          "sizeHuman",
          "headers"
        ],
        "properties": {
          "id": {
            "type": "string",
            "examples": [
              "aB3dEfGh"
            ]
          },
          "time": {
            "type": "string",
            "format": "date-time"
          },
          "read": {
            "type": "boolean"
          },
          "subject": {
            "type": "string"
          },
          "from": {
            "type": [
              "array",
              "null"
            ],
            "items": {
              "$ref": "#/components/schemas/Address"
            }
          },
          "to": {
            "type": [
              "array",
              "null"
            ],
            "items": {
              "$ref": "#/components/schemas/Address"
            }
          },
          "cc": {
            "type": [
              "array",
              "null"
            ],
            "items": {
              "$ref": "#/components/schemas/Address"
            }
          },
          "bcc": {
            "type": [
              "array",
              "null"
            ],
            "items": {
              "$ref": "#/components/schemas/Address"
            }
          },
          "calculatedBcc": {
            "type": [
              "array",
              "null"
            ],
            "items": {
              "$ref": "#/components/schemas/Address"
            }
          },
          "text": {
            "type": "string"
          },
          "html": {
            "type": "string",
            "description": "Sanitized HTML body."
          },
          "attachments": {
            "type": [
              "array",
              "null"
            ],
            "items": {
              "$ref": "#/components/schemas/Attachment"
            }
          },
          "envelope": {
            "$ref": "#/components/schemas/Envelope"
          },
          "source": {
            "type": "string"
          },
          "size": {
            "type": "integer",
            "format": "int64"
          },
          "sizeHuman": {
            "type": "string"
          },
          "headers": {
            "type": [
              "object",
              "null"
            ],
            "additionalProperties": true
          }
        },
        "additionalProperties": false
      },
...
      "APIResponse": {
        "type": "object",
        "required": [
          "code",
          "message"
        ],
        "properties": {
          "code": {
            "type": "string"
          },
          "message": {
            "type": "string"
          },
          "data": {}
        },
        "additionalProperties": false
      },
      "Error": {
        "type": "object",
        "required": [
          "code",
          "error",
          "message"
        ],
        "properties": {
          "code": {
            "type": "string"
          },
          "error": {
            "type": "string"
          },
          "message": {
            "type": "string"
          }
        },
        "additionalProperties": false
      },
...
    },
    "responses": {
      "BadRequest": {
        "description": "The request was rejected.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            },
            "example": {
              "code": "INVALID_REQUEST",
              "error": "INVALID_REQUEST",
              "message": "Invalid request"
            }
          }
        }
      },
      "NotFound": {
        "description": "The email or file was not found.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            },
            "example": {
              "code": "EMAIL_NOT_FOUND",
              "error": "EMAIL_NOT_FOUND",
              "message": "Email not found"
            }
          }
        }
      },
      "Unauthorized": {
        "description": "Basic Auth is enabled and credentials are missing or invalid. The body is plain text, not the JSON API error structure.",
        "headers": {
          "WWW-Authenticate": {
            "schema": {
              "type": "string"
            },
            "description": "Present as `Basic realm=\"OwlMail\"` when the Authorization header is missing."
          }
        },
        "content": {
          "text/plain": {
            "schema": {
              "type": "string"
            },
            "example": "Unauthorized"
          }
        }
      },
      "Forbidden": {
        "description": "When Basic Auth deployment middleware is enabled, the pre-authentication same-origin check rejected a browser request with a foreign `Origin`. A 403 does not imply successful authentication, and the body is plain text.",
        "content": {
          "text/plain": {
            "schema": {
              "type": "string"
            },
            "example": "Forbidden"
          }
        }
      }
    }
  }
}